Former-commit-id: 8bad923169b2a73f249fd8b3153236d6a74100c9

+++ src/rdb.cpp
@@ -36,6 +36,7 @@
 #include "cron.h"
 
 #include <math.h>
+#include <fcntl.h>
 #include <sys/types.h>
 #include <sys/time.h>
 #include <sys/resource.h>
@@ -57,6 +58,9 @@ extern int rdbCheckMode;
 void rdbCheckError(const char *fmt, ...);
 void rdbCheckSetError(const char *fmt, ...);
 
+#ifdef __GNUC__
+void rdbReportError(int corruption_error, int linenum, const char *reason, ...) __attribute__ ((format (printf, 3, 4)));
+#endif
 void rdbReportError(int corruption_error, int linenum, const char *reason, ...) {
     va_list ap;
     char msg[1024];
@@ -85,7 +89,7 @@ void rdbReportError(int corruption_error, int linenum, const char *reason, ...) 
     exit(1);
 }
 
-static int rdbWriteRaw(rio *rdb, void *p, size_t len) {
+static ssize_t rdbWriteRaw(rio *rdb, void *p, size_t len) {
     if (rdb && rioWrite(rdb,p,len) == 0)
         return -1;
     return len;
@@ -496,7 +500,7 @@ void *rdbGenericLoadStringObject(rio *rdb, int flags, size_t *lenptr) {
     int plain = flags & RDB_LOAD_PLAIN;
     int sds = flags & RDB_LOAD_SDS;
     int isencoded;
-    uint64_t len;
+    unsigned long long len;
 
     len = rdbLoadLen(rdb,&isencoded);
     if (isencoded) {
@@ -508,8 +512,8 @@ void *rdbGenericLoadStringObject(rio *rdb, int flags, size_t *lenptr) {
         case RDB_ENC_LZF:
             return rdbLoadLzfStringObject(rdb,flags,lenptr);
         default:
-            rdbExitReportCorruptRDB("Unknown RDB string encoding type %d",len);
-            return nullptr; /* Never reached. */
+            rdbExitReportCorruptRDB("Unknown RDB string encoding type %llu",len);
+            return nullptr;
         }
     }
 
@@ -1219,6 +1223,8 @@ ssize_t rdbSaveSingleModuleAux(rio *rdb, int when, moduleType *mt) {
     /* Save a module-specific aux value. */
     RedisModuleIO io;
     int retval = rdbSaveType(rdb, RDB_OPCODE_MODULE_AUX);
+    if (retval == -1) return -1;
+    io.bytes += retval;
 
     /* Write the "module" identifier as prefix, so that we'll be able
      * to call the right module during loading. */
@@ -1418,7 +1424,7 @@ int rdbSave(const redisDbPersistentDataSnapshot **rgpdb, rdbSaveInfo *rsi)
 int rdbSaveFile(char *filename, const redisDbPersistentDataSnapshot **rgpdb, rdbSaveInfo *rsi) {
     char tmpfile[256];
     char cwd[MAXPATHLEN]; /* Current working dir path for error messages. */
-    FILE *fp;
+    FILE *fp = NULL;
     rio rdb;
     int error = 0;
 
@@ -1447,10 +1453,11 @@ int rdbSaveFile(char *filename, const redisDbPersistentDataSnapshot **rgpdb, rdb
     }
 
     /* Make sure data will not remain on the OS's output buffers */
-    if (fflush(fp) == EOF) goto werr;
-    if (fsync(fileno(fp)) == -1) goto werr;
-    if (fclose(fp) == EOF) goto werr;
-
+    if (fflush(fp)) goto werr;
+    if (fsync(fileno(fp))) goto werr;
+    if (fclose(fp)) { fp = NULL; goto werr; }
+    fp = NULL;
+    
     /* Use RENAME to make sure the DB file is changed atomically only
      * if the generate DB file is ok. */
     if (rename(tmpfile,filename) == -1) {
@@ -1483,7 +1490,7 @@ werr:
         serverLog(LL_WARNING, "Background save cancelled");
     else
         serverLog(LL_WARNING,"Write error saving DB on disk: %s", strerror(errno));
-    fclose(fp);
+    if (fp) fclose(fp);
     unlink(tmpfile);
     stopSaving(0);
     return C_ERR;
@@ -1505,7 +1512,7 @@ void *rdbSaveThread(void *vargs)
 
     int retval = rdbSave(args->rgpdb, &args->rsi);    
     if (retval == C_OK)
-        sendChildCOWInfo(CHILD_INFO_TYPE_RDB, "RDB");
+        sendChildCOWInfo(CHILD_TYPE_RDB, "RDB");
 
     // If we were told to cancel the requesting thread holds the lock for us
     ssize_t cbStart = zmalloc_used_memory();
@@ -1535,7 +1542,7 @@ int rdbSaveBackgroundFork(rdbSaveInfo *rsi) {
     g_pserver->lastbgsave_try = time(NULL);
     openChildInfoPipe();
 
-    if ((childpid = redisFork()) == 0) {
+    if ((childpid = redisFork(CHILD_TYPE_RDB)) == 0) {
         int retval;
 
         /* Child */
@@ -1544,7 +1551,7 @@ int rdbSaveBackgroundFork(rdbSaveInfo *rsi) {
         redisSetCpuAffinity(g_pserver->bgsave_cpulist);
         retval = rdbSave(nullptr, rsi);
         if (retval == C_OK) {
-            sendChildCOWInfo(CHILD_INFO_TYPE_RDB, "RDB");
+            sendChildCOWInfo(CHILD_TYPE_RDB, "RDB");
         }
         exitFromChild((retval == C_OK) ? 0 : 1);
     } else {
@@ -1560,6 +1567,7 @@ int rdbSaveBackgroundFork(rdbSaveInfo *rsi) {
         g_pserver->rdb_save_time_start = time(NULL);
         g_pserver->rdb_child_pid = childpid;
         g_pserver->rdb_child_type = RDB_CHILD_TYPE_DISK;
+        updateDictResizePolicy();
         return C_OK;
     }
     return C_OK; /* unreached */
@@ -1629,11 +1637,33 @@ int rdbSaveBackground(rdbSaveInfo *rsi) {
     return C_OK;
 }
 
-void rdbRemoveTempFile(int tmpfileNum) {
+/* Note that we may call this function in signal handle 'sigShutdownHandler',
+ * so we need guarantee all functions we call are async-signal-safe.
+ * If  we call this function from signal handle, we won't call bg_unlik that
+ * is not async-signal-safe. */
+void rdbRemoveTempFile(int tmpfileNum, int from_signal) {
     char tmpfile[256];
+    char pid[32];
+    char tmpfileNumString[214];
 
-    snprintf(tmpfile,sizeof(tmpfile),"temp-%d-%d.rdb", getpid(), tmpfileNum);
-    unlink(tmpfile);
+    /* Generate temp rdb file name using aync-signal safe functions. */
+    int pid_len = ll2string(pid, sizeof(pid), getpid());
+    int tmpfileNum_len = ll2string(tmpfileNumString, sizeof(tmpfileNumString), tmpfileNum);
+    strcpy(tmpfile, "temp-");
+    strncpy(tmpfile+5, pid, pid_len);
+    strcpy(tmpfile+5+pid_len, "-");
+    strncpy(tmpfile+5+pid_len+1, tmpfileNumString, tmpfileNum_len);
+    strcpy(tmpfile+5+pid_len+1+tmpfileNum_len, ".rdb");
+
+    if (from_signal) {
+        /* bg_unlink is not async-signal-safe, but in this case we don't really
+         * need to close the fd, it'll be released when the process exists. */
+        int fd = open(tmpfile, O_RDONLY|O_NONBLOCK);
+        UNUSED(fd);
+        unlink(tmpfile);
+    } else {
+        bg_unlink(tmpfile);
+    }
 }
 
 /* This function is called by rdbLoadObject() when the code is in RDB-check
@@ -1760,7 +1790,7 @@ robj *rdbLoadObject(int rdbtype, rio *rdb, sds key, uint64_t mvcc_tstamp) {
         zs = (zset*)ptrFromObj(o);
 
         if (zsetlen > DICT_HT_INITIAL_SIZE)
-            dictExpand(zs->pdict,zsetlen);
+            dictExpand(zs->dict,zsetlen);
 
         /* Load every single element of the sorted set. */
         while(zsetlen--) {
@@ -1791,7 +1821,7 @@ robj *rdbLoadObject(int rdbtype, rio *rdb, sds key, uint64_t mvcc_tstamp) {
             if (sdslen(sdsele) > maxelelen) maxelelen = sdslen(sdsele);
 
             znode = zslInsert(zs->zsl,score,sdsele);
-            dictAdd(zs->pdict,sdsele,&znode->score);
+            dictAdd(zs->dict,sdsele,&znode->score);
         }
 
         /* Convert *after* loading, since sorted sets are not stored ordered. */
@@ -2300,8 +2330,11 @@ void stopSaving(int success) {
 void rdbLoadProgressCallback(rio *r, const void *buf, size_t len) {
     if (g_pserver->rdb_checksum)
         rioGenericUpdateChecksum(r, buf, len);
-    if (g_pserver->loading_process_events_interval_bytes &&
-        (r->processed_bytes + len)/g_pserver->loading_process_events_interval_bytes > r->processed_bytes/g_pserver->loading_process_events_interval_bytes)
+    
+    if ((g_pserver->loading_process_events_interval_bytes &&
+        (r->processed_bytes + len)/g_pserver->loading_process_events_interval_bytes > r->processed_bytes/g_pserver->loading_process_events_interval_bytes) ||
+        (g_pserver->loading_process_events_interval_keys &&
+        (r->keys_since_last_callback >= g_pserver->loading_process_events_interval_keys)))
     {
         /* The DB can take some non trivial amount of time to load. Update
          * our cached time since it is used to create and update the last
@@ -2319,6 +2352,14 @@ void rdbLoadProgressCallback(rio *r, const void *buf, size_t len) {
         loadingProgress(r->processed_bytes);
         processEventsWhileBlocked(serverTL - g_pserver->rgthreadvar);
         processModuleLoadingProgressEvent(0);
+
+        robj *ping_argv[1];
+
+        ping_argv[0] = createStringObject("PING",4);
+        replicationFeedSlaves(g_pserver->slaves, g_pserver->replicaseldb, ping_argv, 1);
+        decrRefCount(ping_argv[0]);
+
+        r->keys_since_last_callback = 0;
     }
 }
 
@@ -2423,7 +2464,7 @@ int rdbLoadRio(rio *rdb, int rdbflags, rdbSaveInfo *rsi) {
         } else if (type == RDB_OPCODE_AUX) {
             /* AUX: generic string-string fields. Use to add state to RDB
              * which is backward compatible. Implementations of RDB loading
-             * are requierd to skip AUX fields they don't understand.
+             * are required to skip AUX fields they don't understand.
              *
              * An AUX field is composed of two strings: key and value. */
             robj *auxkey, *auxval;
@@ -2451,7 +2492,7 @@ int rdbLoadRio(rio *rdb, int rdbflags, rdbSaveInfo *rsi) {
                 if (luaCreateFunction(NULL,g_pserver->lua,auxval) == NULL) {
                     rdbExitReportCorruptRDB(
                         "Can't load Lua script from RDB file! "
-                        "BODY: %s", ptrFromObj(auxval));
+                        "BODY: %s", (char*)ptrFromObj(auxval));
                 }
             } else if (!strcasecmp(szFromObj(auxkey),"redis-ver")) {
                 serverLog(LL_NOTICE,"Loading RDB produced by version %s",
@@ -2649,6 +2690,8 @@ int rdbLoadRio(rio *rdb, int rdbflags, rdbSaveInfo *rsi) {
         if (g_pserver->key_load_delay)
             usleep(g_pserver->key_load_delay);
 
+        rdb->keys_since_last_callback++;
+
         /* Reset the state that is key-specified and is populated by
          * opcodes before the key, so that we start from scratch again. */
         expiretime = -1;
@@ -2752,7 +2795,7 @@ int rdbLoadFile(const char *filename, rdbSaveInfo *rsi, int rdbflags) {
 
 /* A background saving child (BGSAVE) terminated its work. Handle this.
  * This function covers the case of actual BGSAVEs. */
-void backgroundSaveDoneHandlerDisk(int exitcode, bool fCancelled) {
+static void backgroundSaveDoneHandlerDisk(int exitcode, bool fCancelled) {
     if (!fCancelled && exitcode == 0) {
         serverLog(LL_NOTICE,
             "Background saving terminated with success");
@@ -2768,23 +2811,16 @@ void backgroundSaveDoneHandlerDisk(int exitcode, bool fCancelled) {
         serverLog(LL_WARNING,
             "Background saving cancelled");
         latencyStartMonitor(latency);
-        rdbRemoveTempFile(g_pserver->rdbThreadVars.tmpfileNum);
+        rdbRemoveTempFile(g_pserver->rdbThreadVars.tmpfileNum, 0);
         latencyEndMonitor(latency);
         latencyAddSampleIfNeeded("rdb-unlink-temp-file",latency);
     }
-    g_pserver->rdbThreadVars.fRdbThreadActive = false;
-    g_pserver->rdb_child_type = RDB_CHILD_TYPE_NONE;
-    g_pserver->rdb_save_time_last = time(NULL)-g_pserver->rdb_save_time_start;
-    g_pserver->rdb_save_time_start = -1;
-    /* Possibly there are slaves waiting for a BGSAVE in order to be served
-     * (the first stage of SYNC is a bulk transfer of dump.rdb) */
-    updateSlavesWaitingBgsave((!fCancelled && exitcode == 0) ? C_OK : C_ERR, RDB_CHILD_TYPE_DISK);
 }
 
 /* A background saving child (BGSAVE) terminated its work. Handle this.
  * This function covers the case of RDB -> Slaves socket transfers for
  * diskless replication. */
-void backgroundSaveDoneHandlerSocket(int exitcode, bool fCancelled) {
+static void backgroundSaveDoneHandlerSocket(int exitcode, bool fCancelled) {
     serverAssert(GlobalLocksAcquired());
 
     if (!fCancelled && exitcode == 0) {
@@ -2796,15 +2832,11 @@ void backgroundSaveDoneHandlerSocket(int exitcode, bool fCancelled) {
         serverLog(LL_WARNING,
             "Background transfer terminated cancelled");
     }
-    g_pserver->rdbThreadVars.fRdbThreadActive = false;
-    g_pserver->rdb_child_type = RDB_CHILD_TYPE_NONE;
-    g_pserver->rdb_save_time_start = -1;
-
-    updateSlavesWaitingBgsave((!fCancelled && exitcode == 0) ? C_OK : C_ERR, RDB_CHILD_TYPE_SOCKET);
 }
 
 /* When a background RDB saving/transfer terminates, call the right handler. */
 void backgroundSaveDoneHandler(int exitcode, bool fCancelled) {
+    int type = g_pserver->rdb_child_type;
     switch(g_pserver->rdb_child_type) {
     case RDB_CHILD_TYPE_DISK:
         backgroundSaveDoneHandlerDisk(exitcode,fCancelled);
@@ -2816,6 +2848,14 @@ void backgroundSaveDoneHandler(int exitcode, bool fCancelled) {
         serverPanic("Unknown RDB child type.");
         break;
     }
+
+    g_pserver->rdbThreadVars.fRdbThreadActive = false;
+    g_pserver->rdb_child_type = RDB_CHILD_TYPE_NONE;
+    g_pserver->rdb_save_time_last = time(NULL)-g_pserver->rdb_save_time_start;
+    g_pserver->rdb_save_time_start = -1;
+    /* Possibly there are slaves waiting for a BGSAVE in order to be served
+     * (the first stage of SYNC is a bulk transfer of dump.rdb) */
+    updateSlavesWaitingBgsave((!fCancelled && exitcode == 0) ? C_OK : C_ERR, type);
 }
 
 /* Kill the RDB saving child using SIGUSR1 (so that the parent will know
@@ -2826,12 +2866,12 @@ void killRDBChild(bool fSynchronous) {
 
     if (cserver.fForkBgSave) {
         kill(g_pserver->rdb_child_pid,SIGUSR1);
-        rdbRemoveTempFile(g_pserver->rdb_child_pid);
+        rdbRemoveTempFile(g_pserver->rdb_child_pid, 0);
         closeChildInfoPipe();
         updateDictResizePolicy();
     } else { 
         g_pserver->rdbThreadVars.fRdbThreadCancel = true;
-        rdbRemoveTempFile(g_pserver->rdbThreadVars.tmpfileNum);
+        rdbRemoveTempFile(g_pserver->rdbThreadVars.tmpfileNum, 0);
         closeChildInfoPipe();
         updateDictResizePolicy();
         if (fSynchronous)
@@ -2840,6 +2880,9 @@ void killRDBChild(bool fSynchronous) {
             serverAssert(!GlobalLocksAcquired());
             void *result;
             int err = pthread_join(g_pserver->rdbThreadVars.rdb_child_thread, &result);
+            if (err) {
+                serverLog(LL_WARNING, "RDB child thread could not be joined: %s", strerror(err));
+            }
             g_pserver->rdbThreadVars.fRdbThreadCancel = false;
             aeAcquireLock();
         }
@@ -2871,7 +2914,7 @@ void *rdbSaveToSlavesSocketsThread(void *vargs)
         retval = C_ERR;
 
     if (retval == C_OK) {
-        sendChildCOWInfo(CHILD_INFO_TYPE_RDB, "RDB");
+        sendChildCOWInfo(CHILD_TYPE_RDB, "RDB");
     }
 
     rioFreeFd(&rdb);
